requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
captcha==0.3
dj-database-url==0.5.0
Django==3.2.10
django-ckeditor==6.2.0
django-cors-headers==3.10.1
django-js-asset==1.2.2
django-recaptcha==2.0.6
gunicorn==20.1.0
Pillow==8.4.0
pkg_resources==0.0.0
psycopg2-binary==2.9.2
pytz==2021.3
sqlparse==0.4.2
whitenoise==5.3.0
